chore: move auth to different file
